fixup! Feat(exporters): Variables exporter export new token structure #DS-1435
